requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1

# ML & Data Science
scikit-learn==1.3.2
xgboost==2.0.2
numpy==1.24.4
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2

# Network Monitoring
scapy==2.5.0
pyshark==0.6
netifaces==0.11.0
psutil==5.9.6

# Security & Encryption
cryptography==41.0.8
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
python-dateutil==2.8.2
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0